# Inplace sums, remove call to `F.pad`  and better memory count