- perform in-place sums of `hidden_states` ie doing `a.add_(b)` instead of `a=a+b` to reduce temp memory allocation
- remove call to asymmetric padding in `F.pad` of non-`replicate` pad mode, and instead let padding be done by `Conv3d` for more efficient execution;
- computation of `memory_count` doesn't extend dimensions to allow `torch.compile` to do a better optimisation (?)