@brunomaga 2025-01-24
Running the following `test_autoencoder.py`

```python
from diffusers.models.autoencoders.autoencoder_kl_cogvideox import CogVideoXCausalConv3d, CogVideoXResnetBlock3D

import time
import torch
import torch.nn as nn


def train(model: nn.Module, video_input: torch.Tensor):
    optimizer = torch.optim.Adam(model.parameters(), lr=0.001)
    loss_fn = torch.nn.MSELoss()
    model.train()
    start_train = time.time()
    for _ in range(100):  # Simulate 100 training iterations
        optimizer.zero_grad()
        output = model(video_input)[0] # returns output, conv_cache
        loss = loss_fn(output*1.1, output)
        loss.backward()
        optimizer.step()
        torch.cuda.synchronize()
    train_time = time.time() - start_train
    print("train_time", train_time, "secs")


def eval(model: nn.Module, video_input: torch.Tensor):
    model.eval()
    start_train = time.time()
    with torch.no_grad():
        for _ in range(300):  # Simulate 300 inference iterations
            output = model(video_input)
            torch.cuda.synchronize()
    eval_time = time.time() - start_train
    print("eval_time", eval_time, "secs")




if __name__=='__main__':
    device='cuda'
    video_input = torch.randn([1, 128, 8, 544, 960], device=device, dtype=torch.bfloat16) # Batch Channels Frames Height Width
    memory_count = torch.prod(torch.tensor(video_input.shape)) / 1024**3
    print("input size: ", memory_count.item(), "GBs")
    model = CogVideoXCausalConv3d(in_channels=video_input.shape[1], out_channels=512, kernel_size=3)
    model = model.to(device=video_input.device, dtype=video_input.dtype)
    torch.cuda.reset_peak_memory_stats()
    eval(model, video_input)
    train(model, video_input)
    print("Max memory", torch.cuda.max_memory_allocated()/ 1024**3, "GBs")
```

calling the main branch gives:

```shell
$ PYTHONPATH=./diffusers_main/src/ python test_autoencoder.py
input size:  0.498046875 GBs
eval_time 33.06385564804077 secs
train_time 34.33984375 secs
Max memory 22.18018913269043 GBs
```

calling this PR branch gives:

```shell
$ PYTHONPATH=./diffusers_PR/src/ python test_autoencoder.py
input size:  0.498046875 GBs
eval_time 31.588099241256714 secs
train_time 34.1251916885376 secs
Max memory 22.17398452758789 GBs
```

I'll try to test more dimensions.