@brunomaga 2025-01-27
@hlky I updated the code above to fix a seed (`torch.manual_seed(42)`) and output the output tensor after 100 training iterations. Then I ran this to compare both `output_*.pt` files:

```python
if __name__=='__main__':
    output_main: torch.Tensor = torch.load("output_main.pt")
    output_PR: torch.Tensor = torch.load("output_PR.pt")
    print(output_main.mean().item(), "vs", output_PR.mean().item())
    print(output_main.std().item(), "vs", output_PR.std().item())
    print("max abs diff", (output_PR-output_main).diff().abs().max().item())
    assert torch.allclose(output_main, output_PR)
```

output:
```
mean: 1.7345882952213287e-08 vs 1.7345882952213287e-08
std: 0.0020904541015625 vs 0.0020904541015625
max abs diff: 0.0
```