@bm-synth 2025-03-09
@hlky  cc @yiyixuxu here's the updated benchmark scritpt that includes inference:

```python

def train(model, inputs, targets, criterion, optimizer):
    model.train()
    optimizer.zero_grad()
    outputs = model(inputs)
    if isinstance(outputs, tuple):
        outputs = outputs[0] # only use the first tensor as output
    loss = criterion(outputs, targets)
    loss.backward()
    optimizer.step()
    return loss.item()

def inference(model, inputs):
    model.eval()
    with torch.no_grad():
        for _ in range(50):
            outputs = model(inputs)
    return outputs

if __name__=='__main__':

    print(f"diffusers version: {diffusers.__version__}")
    to_kw = dict(device="cuda", dtype=torch.bfloat16)

    torch.manual_seed(42)
    model_main = CogVideoXCausalConv3d(in_channels=128, out_channels=512, kernel_size=3).to(**to_kw)
    model_main = torch.compile(model_main, mode="max-autotune-no-cudagraphs")
    optimizer_main = torch.optim.Adam(model_main.parameters(), lr=1e-3)

    torch.manual_seed(42)
    model_pr = CogVideoXCausalConv3d_PR(in_channels=128, out_channels=512, kernel_size=3).to(**to_kw)
    model_pr = torch.compile(model_pr, mode="max-autotune-no-cudagraphs")
    optimizer_pr = torch.optim.Adam(model_pr.parameters(), lr=1e-3)

    criterion = nn.MSELoss()

    # test 5 very different shape inputs
    for shape in (
        (16, 128, 64, 64, 64),
        (8, 128, 128, 32, 32),
        (4, 128, 256, 128, 128),
        (4, 128, 512, 256, 256),
        (2, 128, 1024, 512, 512),
    ):

        input = torch.randn(shape, generator=torch.Generator().manual_seed(0)).to(**to_kw)
        output_shape = ( shape[0], 512, shape[2], shape[3], shape[4] )
        target = torch.randn(output_shape, generator=torch.Generator().manual_seed(0)).to(**to_kw)
        
        # few iterations of training to warm up the model and compiler
        model_main.train()
        model_pr.train()
        for _ in range(5):
            train(model_main, input, target, criterion, optimizer_main)
            train(model_pr, input, target, criterion, optimizer_pr)

        # benchmark training
        runtime_main = benchmark_fn( train, model_main, input, target, criterion, optimizer_main)
        runtime_pr = benchmark_fn( train, model_pr, input, target, criterion, optimizer_pr)
        print(f"shape {shape} train: main runtime {runtime_main} vs PR runtime {runtime_pr}")

        # few iterations of inference to warm up the model and compiler
        torch.compiler.reset()
        model_main.eval()
        model_pr.eval()
        with torch.inference_mode():
            for _ in range(5):
                # few iterations of inference to warm up the model and compiler
                inference(model_main, input)
                inference(model_pr, input)
            runtime_main = benchmark_fn( inference, model_main, input)
            runtime_pr = benchmark_fn( inference, model_pr, input)
            print(f"shape {shape} inference: main runtime {runtime_main} vs PR runtime {runtime_pr}")

        print()
```

to measure memory, i added to the `forward` of `CogVideoXCausalConv3d`:

```python
        print(f"MAIN Memory BEFORE conv:", torch.cuda.memory_allocated())
        output = self.conv(inputs)
        print(f"MAIN Memory AFTER conv:", torch.cuda.memory_allocated())
```

to both PR and main branch. **Memory usage for first shape**:

`main`, train:
- MAIN Memory BEFORE conv: 6595022848
- MAIN Memory AFTER conv: 10889990144

PR, train:
- PR Memory BEFORE conv: 6523719680
- PR Memory AFTER conv: 10818686976

`main`, inference:
- MAIN Memory BEFORE conv: 10916464640
- MAIN Memory AFTER conv: 15211431936

PR, inference:
- PR Memory BEFORE conv: 10845161472
- PR Memory AFTER conv: 15140128768

**Runtimes for first 2 shapes** (NVIDIA H200):

- shape (16, 128, 64, 64, 64) train: main runtime 0.13034 vs PR runtime 0.07435
- shape (16, 128, 64, 64, 64) inference: main runtime 1.49722 vs PR runtime 1.46837
- shape (8, 128, 128, 32, 32) train: main runtime 0.03417 vs PR runtime 0.01869
- shape (8, 128, 128, 32, 32) inference: main runtime 0.38165 vs PR runtime 0.36572

Feel free to run it on your side and correct this benchmark script if you find any issue.